pyproject.toml
[tool.poetry]
name = "hacktjdub"
version = "0.1.0"
description = ""
authors = ["KrishnanShankar <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-mail = "^0.9.1"
python-dotenv = "^0.17.0"
firebase-admin = "^4.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
